refactor: reorganize ProfilingDataBackend interface and related types for clarity in React Profiler
